PPU: Smoother brightness transitions
A real SNES changes the brightness with the DAC in analog so it is
not limited by 5 bits per color.

Flickering colors can be seen in Guile stage in Street Fighter 2
or background in Super Mario Bros during fade in/out.

Increase to 8bit per color to match smooth brightness transitions of
a real SNES.