PPU: Change sprite evaluation
During sprite evaluation the OAM address is being changed. Writes at
that time will go to both OAM and High OAM.

OBJ_RANGE scans through OAM and saves in-range sprite indexes to an
index array.

OBJ_TIME stores the sprite index to be fetched to OBJ_TIME_INDEX. This
index is not changed unless a new sprite is fetched. It uses the index
to read sprite data from OAM.